refactor: rename qnop-app module to qnop-bootstrap
Resolve the confusing phonetic clash between the application layer
(qnop-application: use cases + ports) and the runnable composition root.
The bootstrap module and its package become qnop-bootstrap / io.qnop.bootstrap;
docs, the module map and the ArchUnit layer definition follow.

Refs #1
